feat: Add FlowNode classes; create new graph for global flow_vars
Initial implementation for local/global flow variables. In KNIME, the two key pieces of information are the 'var_name' and the 'default_value'. Global vs. local will behave much the same way, so they are implemented as Nodes that can be added to the main workflow graph (local) or the new global graph (flow_vars).