refactor: Node validation checks parameters
`validation` defined in parent Node class. It checks all options are valid, raising `ParameterValidationError` if not.

Calls to `node.validate()`  are removed from `pyworkflow` when updating/adding node or edge between Nodes. A call to `validate` occurs when the POST route is hit for updating a Node to check any changes to the configuration are valid.